the rest of migration and relation and foreign seted

+++ database/migrations/2020_11_02_201506_create_payments_table.php
@@ -15,8 +15,8 @@ class CreatePaymentsTable extends Migration
     {
         Schema::create('payments', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('order_id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('order_id');
+            $table->unsignedBigInteger('user_id');
             $table->string('ref_number');
             $table->string('res_number');
             $table->string('cart_number');
